Make HmCrypto.PublicKey more generic
