+++ lib/hm_crypto/public_key.ex
@@ -1,8 +1,7 @@
 defmodule HmCrypto.PublicKey do
-  alias HmCrypto.RsaUtils.{RsaPublicKey, RsaPrivateKey}
-  require HmCrypto.RsaUtils.{RsaPublicKey, RsaPrivateKey}
+  import Record
 
-  @type rsa_key :: RsaPublicKey.t() | RsaPrivateKey.t() | binary()
+  @type key :: tuple() | String.t()
 
   @moduledoc """
 
@@ -17,35 +16,11 @@ defmodule HmCrypto.PublicKey do
 
   """
 
-  @spec parse_pem(rsa_key()) :: RsaPublicKey.t() | RsaPrivateKey.t()
+  @spec parse_pem(key()) :: tuple()
   def parse_pem(pem_string) when is_binary(pem_string) and pem_string != "" do
     [pem_entry] = :public_key.pem_decode(pem_string)
     :public_key.pem_entry_decode(pem_entry)
   end
 
-  def parse_pem(RsaPublicKey.record() = pem), do: pem
-  def parse_pem(RsaPrivateKey.record() = pem), do: pem
-
-  @doc """
-
-  Function encodes tuple representation of RSA key to PEM format (string).
-  If string is given as argument - returns it as is.
-
-  """
-
-  @spec encode_pem(rsa_key()) :: String.t()
-  def encode_pem(RsaPublicKey.record() = rsa_key) do
-    to_pem(RsaPublicKey.key_type(), rsa_key)
-  end
-
-  def encode_pem(RsaPrivateKey.record() = rsa_key) do
-    to_pem(RsaPrivateKey.key_type(), rsa_key)
-  end
-
-  def encode_pem(rsa_key) when is_binary(rsa_key), do: rsa_key
-
-  defp to_pem(key_type, rsa_key) do
-    pem_entry = [:public_key.pem_entry_encode(key_type, rsa_key)]
-    :public_key.pem_encode(pem_entry)
-  end
+  def parse_pem(pem) when is_record(pem), do: pem
 end
